backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.22.0
httpx==0.24.0
python-dotenv==1.0.0
pydantic==1.10.23
vercel==0.5.3